third_party/blink/web_tests/platform/generic/fast/forms/select/select-multiple-elements-with-mouse-drag-expected.txt

Select multiple option with mouse drag

P1
P2
P3
P4
P5
P6
P7
PASS document.getElementById("selectId").options[0].selected is true
PASS document.getElementById("selectId").options[1].selected is true
PASS document.getElementById("selectId").options[2].selected is true
PASS document.getElementById("selectId").options[3].selected is true
PASS successfullyParsed is true

TEST COMPLETE
